+++ src/model/Livro.java
@@ -5,7 +5,6 @@ import model.commons.IAvaliavel;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Comparator;
-import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -33,7 +32,7 @@ public class Livro extends Media  implements IAvaliavel {
 
 
     @Override
-    public int getPontuacao() {
+    public double getPontuacao() {
         return avaliacoes.stream()
                 .max(Comparator.comparing(Avaliacao::getDataAvaliacao))
                 .map(Avaliacao::getPontuacao)
