+++ src/model/commons/ITemAvaliacao.java
@@ -1,11 +1,8 @@
 package model.commons;
 
-import model.Avaliacao;
 import model.TipoMedia;
 
-import java.util.List;
-
 public interface ITemAvaliacao extends Identifiable{
-    int getPontuacao();
+    double getPontuacao();
     TipoMedia getTipoMedia();
 }
